[3.0-pre10] Fix file listener and command.
