+++ build.gradle.kts
@@ -47,7 +47,8 @@ taboolib {
 
     relocate("net.wesjd.anvilgui", "${project.group}.module.internal.inputer.anvil")
 
-    version = "6.0.0-pre18"
+    classifier = null
+    version = "6.0.0-pre19"
 }
