--- src/main/kotlin/me/arasple/mc/trmenu/module/internal/command/impl/CommandExpresser.kt
@@ -1,9 +1,0 @@
-package me.arasple.mc.trmenu.module.internal.command.impl
-
-import taboolib.common.platform.SimpleCommandBody
-
-interface CommandExpresser {
-
-    val command: SimpleCommandBody
-
-}
