+++ src/main/kotlin/me/arasple/mc/trmenu/module/internal/command/impl/CommandTemplate.kt
@@ -23,45 +23,11 @@ import taboolib.platform.util.sendLang
  * @author Arasple
  * @date 2020/7/22 12:08
  */
-object CommandTemplate : CommandExpresser {
-
-    override val command = subCommand {
-        dynamic(optional = true) {
-            suggestion<CommandSender> { _, _ ->
-                listOf("1", "2", "3", "4", "5", "6")
-            }
-        }
-        execute<Player> { player, context, argument ->
-            val rows = (if (context.args.isNotEmpty()) NumberUtils.toInt(context.args[0], 5) else 3).coerceAtMost(6)
-
-            buildMenu<Basic>("Template#$rows") {
-                rows(rows)
-                handLocked(false)
-                onClose { e ->
-                    val inventory = e.inventory
-
-                    if (inventory.all { it == null || it.type == Material.AIR }) {
-                        player.sendLang("Command-Template-Empty")
-                        return@onClose
-                    }
-
-                    XSound.BLOCK_NOTE_BLOCK_BIT.play(player, 1f, 0f)
-                    Paster.paste(player, generate(inventory), "yml")
-
-                    inventory.contents.forEach {
-                        if (!(it == null || it.type == Material.AIR)) {
-                            player.inventory.addItem(it).values.forEach { e -> player.world.dropItem(player.location, e) }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
+object CommandTemplate {
     /**
      * 生成模板
      */
-    private fun generate(inventory: Inventory): String {
+    internal fun generate(inventory: Inventory): String {
         val rows = inventory.size / 9
         val keys = MenuLayout.commonKeys.iterator()
         val items = collectItems(inventory)
